external/virtualization-layer: nvidia-container-toolkit: install config.toml symlink
With the updated toolkit/runtime setup, the toolkit is run at boot time to generate the
configuration file used by the container runtime.  To allow for a read-only rootfs setup,
we place the generated config.toml under /run. We need to install a symlink in the
normal /etc/nvidia-container-runtime directory to point to that generated config, so the
container runtime can find it.

Signed-off-by: Matt Madison <matt@madison.systems>